# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Title",
        "description": "Your API description",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/login": {
            "post": {
                "summary": "Authenticate user",
                "operationId": "2dd578ff7aba2721293dac66833b27d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Register a new user",
                "operationId": "0903f2e0df1a8857bee538fc055417f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/expense-categories": {
            "get": {
                "summary": "Get a list of expense categories",
                "operationId": "3a57961c189fa263f2f2ee032139f434",
                "responses": {
                    "200": {
                        "description": "List of expense categories"
                    }
                }
            },
            "post": {
                "summary": "Create a new expense category",
                "operationId": "e9c91d5e2ba41405bebcb5f3781cb86a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense category created"
                    },
                    "400": {
                        "description": "Expense category already exists"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expense-categories/{id}": {
            "get": {
                "summary": "Get details of a specific expense category",
                "operationId": "0715450f6ef6057c18e5a718092b64be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense category details"
                    },
                    "404": {
                        "description": "Expense category not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific expense category",
                "operationId": "b63dae36bfd9d298798693cad380466e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense category updated"
                    },
                    "404": {
                        "description": "Expense category not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific expense category",
                "operationId": "3ede5ddac795968a7e80a1e4494435db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense category deleted"
                    },
                    "404": {
                        "description": "Expense category not found"
                    }
                }
            }
        },
        "/expenses": {
            "get": {
                "summary": "Get a list of expenses",
                "operationId": "36e62ecfbcfde37686ea63fe89d0c0c0",
                "responses": {
                    "200": {
                        "description": "List of expenses"
                    }
                }
            },
            "post": {
                "summary": "Create a new expense",
                "operationId": "927e29ac31ae983af0a55ce5d487d554",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "property1": {
                                        "type": "type1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Expense created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/expenses/{id}": {
            "get": {
                "summary": "Get details of a specific expense",
                "operationId": "b8d5c673ecdbb6a51bea7ec7cc5d0e30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Expense details"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            },
            "put": {
                "summary": "Update a specific expense",
                "operationId": "36a4200b139cad8582ff017bbdb5750e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "property1": {
                                        "type": "type1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Expense updated"
                    },
                    "404": {
                        "description": "Expense not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "summary": "Delete a specific expense",
                "operationId": "b261a1443844b958952ae83b03fbbfa3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the expense",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Expense deleted"
                    },
                    "404": {
                        "description": "Expense not found"
                    }
                }
            }
        },
...
        "/wallet": {
            "get": {
                "summary": "Get wallet information",
                "operationId": "92884b373e354bacec5be74f2e046805",
                "responses": {
                    "200": {
                        "description": "Wallet information"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    }
}